fix(invoker): drop a redundant uid cast that reds clippy on linux
`uid_t`/`gid_t` are `u32` on both supported unix platforms, so the cast was
`u32 -> u32` and `clippy::unnecessary_cast` (CI runs `-D warnings`) rejected it.
Only visible on a unix host: the code is `cfg(unix)`, so a clippy run on Windows
never lints it.

+++ src/invoker.rs
@@ -379,9 +379,12 @@ unsafe fn entry_from_passwd(pwd: &libc::passwd) -> Option<PasswdEntry> {
         return None;
     }
     Some(PasswdEntry {
+        // No cast: `uid_t`/`gid_t` are `u32` on both supported unix platforms, so a cast would be
+        // redundant (and `clippy::unnecessary_cast` rejects it). A platform where they widened would
+        // fail to compile here, which is the right way to find out.
         name,
-        uid: pwd.pw_uid as u32,
-        gid: pwd.pw_gid as u32,
+        uid: pwd.pw_uid,
+        gid: pwd.pw_gid,
         home: PathBuf::from(home),
     })
 }
